pyproject.toml
[project]
name = "webvory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi==0.104.1",
    "httpx==0.25.1",
    "numpy>=1.26.4",
    "openai==1.3.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pytest==7.4.3",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "streamlit==1.28.0",
    "uvicorn==0.24.0",
]
